Revert "Disable weak ciphers via configure"
This reverts commit ae5de255833cdabe0dc69cd9fb190d217038aad1.

This was meant for openvswitch ciphers, which led to the issues with XO
SDN Controller plugin not being able to connect to OVS anymore, the
change in the openvswitch package was reverted, and this one is
therefore not needed.

Note: this does not impact the connexions to XAPI, as the ciphers list
is fixed in stunnel configuration.

Signed-off-by: David Morel <david.morel@vates.tech>